requirements.txt
spotipy==2.23.0
applemusicpy==0.3.5
Pillow==10.2.0
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
loguru==0.7.2 